pyproject.toml
[project]
name = "pizza-order-voice-agent"
version = "0.1.0"
description = "Autonomous voice agent that places outbound pizza delivery calls"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "twilio>=9.0.0",
    "deepgram-sdk>=3.5.0",
    "cartesia>=1.0.0",
    "groq>=0.9.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
